requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchvision
torchaudio
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
timm==0.6.7
ftfy
regex
einops
fvcore
decord==0.6.0
python-dotenv
deeplake
gradio